(nvim) Disable extending comment on 'o' and 'O'
When the cursor is on a source code comment, and 'o' or 'O' is used to
insert a new line, vim can auto insert comment on the new line as well.
This commit disables that option.

Note that pressing Enter at the end of a comment line in insert mode
will still insert comment.